# docs/api-spec.yaml
openapi: 3.0.0
info:
  title: Prompt Manager API
  version: 0.1.0
  description: API for managing prompts, schemas, and test suites
paths:
  /api/v1/prompts:
    get:
      summary: List prompts
      parameters:
        - name: skip
          in: query
          schema: { type: integer, default: 0 }
        - name: limit
          in: query
          schema: { type: integer, default: 100 }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Prompt' }
    post:
      summary: Create prompt
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PromptCreate' }
      responses:
        '201': { description: Created }
  # ... other endpoints similar
  /api/v1/admin/api-keys:
    post:
      summary: Create LLM API key
      parameters:
        - name: org_id
          in: query
          required: true
          schema: { type: string, format: uuid }
        - name: provider
          in: query
          required: true
          schema: { type: string }
        - name: name
          in: query
          required: true
          schema: { type: string }
        - name: value
          in: query
          required: true
          schema: { type: string }
      responses:
        '200': { description: OK }
    get:
      summary: List LLM API keys
      parameters:
        - name: org_id
          in: query
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200': { description: OK }
  /api/v1/admin/api-keys/{key_id}/rotate:
    post:
      summary: Rotate a stored LLM API key
      parameters:
        - name: key_id
          in: path
          required: true
          schema: { type: string, format: uuid }
        - name: value
          in: query
          required: true
          schema: { type: string }
      responses:
        '200': { description: OK }
  /api/v1/admin/api-keys/{key_id}/reveal:
    post:
      summary: Reveal a stored LLM API key
      parameters:
        - name: key_id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200': { description: OK }
  /api/v1/admin/api-keys/rewrap:
    post:
      summary: Rewrap stored LLM API keys after backend rotation
      parameters:
        - name: org_id
          in: query
          required: false
          schema: { type: string, format: uuid }
      responses:
        '200': { description: OK }
  /api/v1/admin/secrets/rotate:
    post:
      summary: Rotate the configured secrets backend (Vault Transit)
      responses:
        '200': { description: OK }
  /api/v1/internal/orgs/{org_id}/api-keys/{provider}:
    get:
      summary: Resolve an org-scoped API key (internal)
      parameters:
        - name: org_id
          in: path
          required: true
          schema: { type: string, format: uuid }
        - name: provider
          in: path
          required: true
          schema: { type: string }
        - name: name
          in: query
          required: false
          schema: { type: string }
      responses:
        '200': { description: OK }
components:
  schemas:
    Prompt:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        template: { type: string }
        description: { type: string }
        version: { type: string }
        metadata: { type: object }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    PromptCreate:
      type: object
      required: [name, template]
      properties:
        name: { type: string }
        template: { type: string }
        description: { type: string }
        version: { type: string, default: "1.0.0" }
        metadata: { type: object }
    # ... etc.